VCenterServerForm: check delete button only if any
fixes #178

+++ library/Vspheredb/Web/Form/VCenterServerForm.php
@@ -160,7 +160,7 @@ class VCenterServerForm extends Form
             $this->registerElement($button);
         }
         $this->add($buttons);
-        if ($deleteButton->hasBeenPressed()) {
+        if ($deleteButton && $deleteButton->hasBeenPressed()) {
             $this->getObject()->delete();
             $this->deleted = true;
         }
